pyproject.toml
[project]
name = "canary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastparquet>=2024.11.0",
    "fsspec>=2024.12.0",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "onnx>=1.17.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
keras = [
    "keras>=3.7.0",
    "tensorflow>=2.16.0",
    "tensorflow-macos>=2.16.0 ; platform_machine == 'arm64' and sys_platform == 'darwin'",
    "tensorflow-metal>=1.1.0 ; platform_machine == 'arm64' and sys_platform == 'darwin'",
    "tf2onnx>=1.16.1",
]
torch = [
    "onnxscript>=0.1.0.dev20250108",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
]
albumentations = [
    "albumentations>=2.0.0",
]
xgboost = [
    "onnxmltools>=1.13.0",
    "xgboost>=2.1.3",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "python-semantic-release>=9.15.2",
]
